+++ src/tewi/service/search/__init__.py
@@ -1,5 +1,7 @@
 from .base_provider import BaseSearchProvider
 from .yts_provider import YTSProvider
 from .torrentscsv_provider import TorrentsCsvProvider
+from .tpb_provider import TPBProvider
 
-__all__ = ['BaseSearchProvider', 'YTSProvider', 'TorrentsCsvProvider']
+__all__ = ['BaseSearchProvider', 'YTSProvider', 'TorrentsCsvProvider',
+           'TPBProvider']